requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.12
pillow==10.4.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
